io_uring: fix: flatten threads for try_join! macro
The try join was not working as supposed to:
"Waits on multiple concurrent branches, returning when all branches
complete with Ok(_) or on the first Err(_)."
But if a single thread return with Err, try_join was still running.

The original code was from this example:
https://fasterthanli.me/articles/remote-development-with-rust-on-fly-io?ref=dailydev#a-wonderful-tcp-proxy-with-tokio-uring

But the threads was not flattened like an official tokio docs is saying:
https://docs.rs/tokio/latest/tokio/macro.try_join.html#examples
(second example: "Using try_join! with spawned tasks.")

This commit is fixing this problem.