bluetooth: fix a deadlock when phone doesn't connect to AA profile
Await-ing on `handle_aa` should return a handle or should time out,
but there is a problem that it sometimes stay at this almost forever...
It happened to me at least twice - so indeed this is a real issue.

Add a timeout around this call, so we would always handle this properly.