config: remove `keepalive` option for Bluetooth
The `keepalive` option is removed because it has little effect.
Bluetooth profile announcing and the agent are disabled anyway.

Other applications may run concurrently to access data via Bluetooth,
so it's better not to interfere by toggling Bluetooth on/off.

Additionally, disabling this option does not reduce power consumption,
as the device uses a single radio chip.